refactor: remove PublicLessons and change it with LessonsList
