+++ front/src/pages/User/Home/OngoingLessons/OngoingLessons.jsx
@@ -5,7 +5,7 @@ import { USER_ENROLLED_SHORT_LESSONS_BASE_KEY } from '@sb-ui/utils/queries';
 import { getEnrolledLessons } from '@sb-ui/utils/api/v1/student';
 import OngoingLessonsMobile from './OngoingLessons.mobile';
 import OngoingLessonsDesktop from './OngoingLessons.desktop';
-import { PAGE_SHORT_SIZE } from '../PublicLessons/constants';
+import { PAGE_SHORT_SIZE } from './constants';
 
 const OngoingLessons = () => {
   const isMobile = useContext(MobileContext);
