--- front/src/pages/User/Home/PublicLessons/constants.js
@@ -1,2 +1,0 @@
-export const PAGE_SIZE = 10;
-export const PAGE_SHORT_SIZE = 3;
