--- front/src/pages/User/Home/PublicLessons/index.js
@@ -1,3 +1,0 @@
-import PublicLessons from './PublicLessons';
-
-export default PublicLessons;
