+++ front/src/pages/User/Lessons/LessonsList/index.js
@@ -1,0 +1,3 @@
+import LessonsList from './LessonsList';
+
+export default LessonsList;
