--- front/src/pages/User/Lessons/UserLessonsList/index.js
@@ -1,3 +1,0 @@
-import UserLessonsList from './UserLessonsList';
-
-export default UserLessonsList;
